Add files via upload
The project entitled File MANAGEMENT SYSTEM is a web application to create software for managing files by providing a web solution for them. The web solution provides facility for manipulating the internal operation of the firm.
The first problem is that there are loads of hard copied documents are being
generated. Keeping the information in the form of hard copied documents lead to many problems. All the process done manually at the centers and all the records are maintained on the papers. So, the maintenance of the record is very difficult in the departments as well as it is very difficult for the staff to check the record.
The existing system is monotonous, time consuming, less flexible and provides a very hectic working schedule. The chance of loss of record is very high and record searching is very difficult. Maintenance of the system is also very difficult and take a lot of time. Result processing is slow due to paperwork and requirement of staff.
File Management System is an efficient, time saving and easy way to report, view and control the version of a file. It is now an easy task and managing it is much easier. The software should be designed and developed in such a way that it accommodates all needs in dynamic fashion. It should be a general purpose one.